move offlineevalconfig to eval section
